# server/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "virtu API",
    "description": "SimpleLogin-compatible email alias API. Authenticate with the `Authentication: <api_key>` header (SimpleLogin's exact header name — not Authorization). Errors use the {\"error\": \"...\"} envelope.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authentication",
        "description": "API key obtained from POST /auth/login"
      }
    },
    "schemas": {
...
      "CreateMailboxRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "example": {
          "email": "second@example.com"
        }
      },
      "VerifyMailboxRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "example": {
          "code": "662302"
        }
      },
      "UpdateMailboxRequest": {
        "type": "object",
        "properties": {
          "default": {
            "type": "boolean"
          },
          "email": {
            "type": "string"
          },
          "cancel_email_change": {
            "type": "boolean"
          },
          "trash": {
            "type": "boolean"
          }
        }
      },
      "DeleteMailboxRequest": {
        "type": "object",
        "properties": {
          "transfer_aliases_to": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "description": "SimpleLogin-compatible error envelope"
      },
...
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": false
      },
      "DeletedResponse": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ],
        "additionalProperties": false
      },
...
      "MailboxesResponse": {
        "type": "object",
        "properties": {
          "mailboxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Mailbox"
            }
          }
        },
        "required": [
          "mailboxes"
        ],
        "additionalProperties": false
      },
      "Mailbox": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "email": {
            "type": "string"
          },
          "verified": {
            "type": "boolean"
          },
          "default": {
            "type": "boolean"
          },
          "creation_timestamp": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "nb_alias": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "trash": {
            "type": "boolean"
          },
          "suppressed": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "verified",
          "default",
          "creation_timestamp",
          "nb_alias",
          "trash",
          "suppressed"
        ],
        "additionalProperties": false
      },
      "UpdatedResponse": {
        "type": "object",
        "properties": {
          "updated": {
            "type": "boolean"
          }
        },
        "required": [
          "updated"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/v2/mailboxes": {
      "get": {
        "tags": [
          "Mailbox"
        ],
        "description": "Get the user's mailboxes, including unverified ones.",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailboxesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/mailboxes": {
      "post": {
        "tags": [
          "Mailbox"
        ],
        "description": "Create a new mailbox. It starts unverified: a 6-digit code (15-minute expiry) is emailed to the address; confirm it via POST /mailboxes/{mailbox_id}/verify.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMailboxRequest"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mailbox"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/mailboxes/{mailbox_id}/verify": {
      "post": {
        "tags": [
          "Mailbox"
        ],
        "description": "Verify a mailbox with the emailed 6-digit code. Deviation from SimpleLogin (which verifies through a web link): the API accepts the code directly. Error strings follow mailbox_utils.py; 410 once the code has been tried wrongly too many times (re-create the mailbox to get a fresh code). Verifying an already-verified mailbox is a no-op that returns it.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMailboxRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "mailbox_id",
            "required": true
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mailbox"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/mailboxes/{mailbox_id}/verify/request": {
      "post": {
        "tags": [
          "Mailbox"
        ],
        "description": "Email a fresh 6-digit verification code for a mailbox that is unverified or bounce-suppressed (suppressed mailboxes resume forwarding only after re-verifying — proof the address receives mail again). 400 when the mailbox is already verified and not suppressed.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "mailbox_id",
            "required": true
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/mailboxes/{mailbox_id}": {
      "put": {
        "tags": [
          "Mailbox"
        ],
        "description": "Update a mailbox. `default` sets the default mailbox; `trash` designates (true) or clears (false) it as the account's trash inbox, where mail for disabled aliases is delivered. `email` change returns 400 (not implemented yet) and `cancel_email_change` is a no-op.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMailboxRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "mailbox_id",
            "required": true
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Mailbox"
        ],
        "description": "Delete a mailbox. Its aliases are deleted (and tombstoned) unless transfer_aliases_to names another of your verified mailboxes (-1 = delete).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "x-fastify-zod-openapi-optional": true,
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DeleteMailboxRequest"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "mailbox_id",
            "required": true
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api",
      "description": "API root"
    }
  ],
...
}